Add day 9 of year 2019

+++ src/y2019/d02.rs
@@ -4,7 +4,7 @@ pub fn parse_input(input: &str) -> Program {
 	Program::from(input)
 }
 
-pub fn part1(input: &Program) -> i32 {
+pub fn part1(input: &Program) -> i64 {
 	let mut interpreter = Interpreter::from(input);
 	interpreter.ram[1] = 12;
 	interpreter.ram[2] = 2;
@@ -12,7 +12,7 @@ pub fn part1(input: &Program) -> i32 {
 	interpreter.ram[0]
 }
 
-pub fn part2(input: &Program) -> i32 {
+pub fn part2(input: &Program) -> i64 {
 	for noun in 0..100 {
 		for verb in 0..100 {
 			let mut interpreter = Interpreter::from(input);
